+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed auth bypass for `/health` and `/metrics` endpoints.
+  Those endpoints should be always excluded from authentication but that was
+  broken in `v0.73` #2465.
+
 ### Added
 
 - `listen:tls:cert` and `listen:tls:key` config options for listening on HTTPS